Consolidate RL schedulers and translator stub
