+++ src/streaming_dataset_watcher.py
@@ -16,17 +16,7 @@ from .cross_lingual_fairness import CrossLingualFairnessEvaluator
 try:  # pragma: no cover - optional dependency
     from .data_ingest import CrossLingualTranslator
 except Exception:  # pragma: no cover - missing torch
-    class CrossLingualTranslator:  # type: ignore
-        def __init__(self, languages):
-            self.languages = list(languages)
-
-        def translate(self, text: str, lang: str) -> str:
-            if lang not in self.languages:
-                raise ValueError("unsupported language")
-            return text
-
-        def translate_all(self, text: str):
-            return {l: text for l in self.languages}
+    from .translator_fallback import CrossLingualTranslator
 
 
 class StreamingDatasetWatcher:
